[REFACTOR] 챌린지 인증글과 좋아요 엔티티 패키지 변경 (#13)
* fix: Book import 수정

* refactor: 챌린지 인증글 패키지 변경

* refactor: 좋아요 패키지 변경

* style: 공백 및 import 구문 포맷팅

* refactor: 필드명 카멜케이스로 변경

* fix: Wishlist 엔티티 생성자 접근지정자 protected로 변경

생성자 호출을 통한 Wishlist 인스턴스 생성을 방지

* style: 불필요한 공백 제거

* docs: PR 템플릿 리팩토링 필드 추가
