+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -2,6 +2,7 @@
 - [ ] 기능 추가
 - [ ] 기능 삭제
 - [ ] 버그 수정
+- [ ] 리팩토링
 - [ ] 의존성, 환경 변수, 빌드 관련 코드 업데이트
 
 ## 작업 내용 및 변경 사항
