+++ src/main/java/kakao/rebit/book/repository/BookRepository.java
@@ -4,4 +4,5 @@ import kakao.rebit.book.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BookRepository extends JpaRepository<Book, Long> {
+
 }
