+++ src/main/java/kakao/rebit/challenge/repository/ChallengeCertificationRepository.java
@@ -1,6 +1,6 @@
-package kakao.rebit.challengecertification.repository;
+package kakao.rebit.challenge.repository;
 
-import kakao.rebit.challengecertification.entity.ChallengeCertification;
+import kakao.rebit.challenge.entity.ChallengeCertification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChallengeCertificationRepository extends JpaRepository<ChallengeCertification, Long> {
