+++ src/main/java/kakao/rebit/diary/repository/DiaryRepository.java
@@ -4,4 +4,5 @@ import kakao.rebit.diary.entity.Diary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DiaryRepository extends JpaRepository<Diary, Long> {
+
 }
