+++ src/main/java/kakao/rebit/feed/entity/Likes.java
@@ -1,4 +1,4 @@
-package kakao.rebit.Likes.entity;
+package kakao.rebit.feed.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -8,7 +8,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import kakao.rebit.feed.entity.Feed;
 import kakao.rebit.member.entity.Member;
 
 @Entity
@@ -28,7 +27,6 @@ public class Likes {
     private Feed feed;
 
     protected Likes() {
-
     }
 
     public Likes(Member member, Feed feed) {
