+++ src/main/java/kakao/rebit/wishlist/entity/BookWishlist.java
@@ -6,7 +6,7 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import java.awt.print.Book;
+import kakao.rebit.book.entity.Book;
 import kakao.rebit.member.entity.Member;
 
 @Entity
@@ -30,4 +30,3 @@ public class BookWishlist extends Wishlist {
         return book;
     }
 }
-
