+++ src/main/java/kakao/rebit/wishlist/entity/Wishlist.java
@@ -2,6 +2,7 @@ package kakao.rebit.wishlist.entity;
 
 import jakarta.persistence.DiscriminatorColumn;
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -9,7 +10,6 @@ import jakarta.persistence.Inheritance;
 import jakarta.persistence.InheritanceType;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.Table;
 import kakao.rebit.common.persistence.BaseEntity;
 import kakao.rebit.member.entity.Member;
@@ -31,7 +31,7 @@ public class Wishlist extends BaseEntity {
     protected Wishlist() {
     }
 
-    public Wishlist(Member member) {
+    protected Wishlist(Member member) {
         this.member = member;
     }
